Переработка наследования виджетов
Теперь для элементов интерфейсов есть класс Widget, включающий в себя базовую реализацию общих для виджетов параметров и методов, а рисование и взаимодействие переложено на миксины, которые можно передать любым другим новым классам, требующим визуального оформления или взаимодействия
