+++ bin/src/levels/menu.dart
@@ -13,10 +13,10 @@ class Menu extends Level {
     super.board.add(Button(
         Point((width / 2).toInt() + 20, (height / 2).toInt() - (botWidth/2).toInt()), "Exit", width: botWidth, height: botHeig));
   }
-  
+
   @override
   bool get isGameOver => false;
-
+  
   @override
   void timerUpdate() {
     return;
