+++ bin/src/mixins/mdrawable.dart
@@ -1,0 +1,13 @@
+import '../point.dart';
+
+import "../canvas.dart";
+
+/// Интерфейс отрисовываемых объектов.
+///
+/// Объект хранит ифнормацию о то, что нужно отрисовать и передаёт её холсту (отрисовщику)
+mixin MDrawable {
+
+  /// Отрисовка объекта при помощи холста
+  void draw(Canvas engine);
+  
+}
